requirements.txt
flask
flask-cors
azure-storage-blob
azure-ai-formrecognizer
python-dotenv
python-docx
azure-functions
azure-core